Fixed new card cell working
Added sections to collection view to be able to show fixed cell at
beginning of collectionView. We don’t need to modify the saved card
input to indexPath.row 1 because it’s embedded into the second section
of the collectionView.

Need to wire up transition and clean up the offsets.

Also combined all cells into a group.